+++ dune-server/src/main/java/lv/k2611a/domain/ConstructionOption.java
@@ -5,32 +5,32 @@ import java.util.Set;
 
 public enum ConstructionOption {
 
-    DEVASTATOR(1,UnitType.DEVASTATOR),
-    SIEGE_TANK(2,UnitType.SIEGE_TANK),
-    TANK(3,UnitType.SIEGE_TANK),
-    LAUNCHER(4,UnitType.LAUNCHER),
+    DEVASTATOR((byte)1,UnitType.DEVASTATOR),
+    SIEGE_TANK((byte)2,UnitType.SIEGE_TANK),
+    TANK((byte)3,UnitType.SIEGE_TANK),
+    LAUNCHER((byte)4,UnitType.LAUNCHER),
 
-    POWERPLANT(5,BuildingType.POWERPLANT),
-    HARVESTER(6,UnitType.HARVESTER),
-    RADAR(7,BuildingType.RADAR),
-    CONCRETE(8, BuildingType.CONCRETE),
-    JEEP(9, UnitType.JEEP),
-    DEVIATOR(10, UnitType.DEVIATOR),
-    REFINERY(11, BuildingType.REFINERY),
-    FACTORY(12,BuildingType.FACTORY),
-    ROCKET_TURRET(13, BuildingType.ROCKET_TURRET),
-    SILO(14,BuildingType.SILO),
-    REPAIRSHOP(15,BuildingType.REPAIRSHOP),
-    AIRBASE(16,BuildingType.AIRBASE),
-    WALL(17, BuildingType.WALL),
-    MCV(18, UnitType.MCV),
-    TURRET(19, BuildingType.TURRET),
-    SONIC_TANK(20, UnitType.SONIC_TANK),
-    TRIKE(21, UnitType.TRIKE);
+    POWERPLANT((byte)5,BuildingType.POWERPLANT),
+    HARVESTER((byte)6,UnitType.HARVESTER),
+    RADAR((byte)7,BuildingType.RADAR),
+    CONCRETE((byte)8, BuildingType.CONCRETE),
+    JEEP((byte)9, UnitType.JEEP),
+    DEVIATOR((byte)10, UnitType.DEVIATOR),
+    REFINERY((byte)11, BuildingType.REFINERY),
+    FACTORY((byte)12,BuildingType.FACTORY),
+    ROCKET_TURRET((byte)13, BuildingType.ROCKET_TURRET),
+    SILO((byte)14,BuildingType.SILO),
+    REPAIRSHOP((byte)15,BuildingType.REPAIRSHOP),
+    AIRBASE((byte)16,BuildingType.AIRBASE),
+    WALL((byte)17, BuildingType.WALL),
+    MCV((byte)18, UnitType.MCV),
+    TURRET((byte)19, BuildingType.TURRET),
+    SONIC_TANK((byte)20, UnitType.SONIC_TANK),
+    TRIKE((byte)21, UnitType.TRIKE);
 
 
     static {
-        Set<Integer> idsOnJs = new HashSet<Integer>();
+        Set<Byte> idsOnJs = new HashSet<Byte>();
         for (ConstructionOption buildingType : values()) {
             if (!idsOnJs.add(buildingType.getIdOnJS())) {
                 throw new AssertionError("Duplicate js id" + idsOnJs);
@@ -38,15 +38,15 @@ public enum ConstructionOption {
         }
     }
 
-    private final int idOnJS;
+    private final byte idOnJS;
     private final EntityType entityType;
 
-    private ConstructionOption(int idOnJS, EntityType entityType) {
+    private ConstructionOption(byte idOnJS, EntityType entityType) {
         this.idOnJS = idOnJS;
         this.entityType = entityType;
     }
 
-    public int getIdOnJS() {
+    public byte getIdOnJS() {
         return idOnJS;
     }
